backend/use-cases/src/main/kotlin/com/y9vad9/pomodoro/backend/usecases/timers/invites/JoinByInviteUseCase.kt
package com.y9vad9.pomodoro.backend.usecases.timers.invites

import com.y9vad9.pomodoro.backend.repositories.TimerInvitesRepository
import com.y9vad9.pomodoro.backend.repositories.TimersRepository
import com.y9vad9.pomodoro.backend.repositories.UsersRepository

class JoinByInviteUseCase(
    private val invites: TimerInvitesRepository,
    private val timers: TimersRepository
) {
    suspend operator fun invoke(
        userId: UsersRepository.UserId,
        code: TimerInvitesRepository.Code
    ): Result {
        val invite = invites.getInvite(code) ?: return Result.NotFound
        timers.addMember(userId, invite.timerId)

        if (invite.limit.int <= 1)
            invites.removeInvite(invite.code)
        else invites.setInviteLimit(
            code,
            TimerInvitesRepository.Limit(invite.limit.int - 1)
        )

        return Result.Success(invite.timerId)
    }

    sealed interface Result {
        @JvmInline
        value class Success(val timerId: TimersRepository.TimerId) : Result
        object NotFound : Result
    }
}